UCP/RNDV: Move common rndv files to proto dir
